update endpoint documentation and reorganise tests into separate files to be required in to app.test.js
